Include ListMovieCarousel and listConfig to have separate rendering logic
