+++ src/components/Carousel/carousel.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Carousel } from "react-responsive-carousel";
 import "react-responsive-carousel/lib/styles/carousel.min.css";
-import useMoviesWithTrailers from './MoviesTrailersCarousel';
+import useFetchMovies from './ListMoviesCarousel'; 
 
 const CarouselReactResponsive = () => {
-  const movies = useMoviesWithTrailers();
+  const movies = useFetchMovies();
 
   return (
-    <div className="bg-black p-4">
+    <div style={{ backgroundColor: 'black', padding: '16px' }}>
       <Carousel
         showArrows={true}
         showThumbs={false}
@@ -17,25 +17,16 @@ const CarouselReactResponsive = () => {
         className="text-white"
       >
         {movies.map(movie => (
-          <div key={movie.id} className="relative flex flex-col justify-center items-center w-full" style={{ height: '500px' }}>
-            <img src={`https://image.tmdb.org/t/p/w1280${movie.backdrop_path}`} alt={movie.title || movie.name} className="shadow-lg object-cover w-full h-full" style={{ objectFit: 'cover' }} />
-            <div className="absolute bottom-4 left-4 bg-black bg-opacity-50 p-4 rounded">
-              <h3 className="text-xl font-bold">{movie.title || movie.name}</h3>
-              <p className="text-sm">{movie.overview}</p>
+          <div key={movie.id} style={{ position: 'relative', height: '500px' }}>
+            <img 
+              src={`https://image.tmdb.org/t/p/w1280${movie.backdrop_path}`} 
+              alt={movie.title || movie.name} 
+              style={{ width: '100%', height: '100%', objectFit: 'cover', objectPosition: 'center' }} 
+            />
+            <div style={{ position: 'absolute', bottom: '16px', left: '16px', backgroundColor: 'rgba(0, 0, 0, 0.5)', padding: '16px', borderRadius: '8px' }}>
+              <h3 style={{ fontSize: '1.5rem', fontWeight: 'bold' }}>{movie.title || movie.name}</h3>
+              <p style={{ fontSize: '1rem' }}>{movie.overview}</p>
             </div>
-            {movie.trailer && (
-              <div className="absolute inset-0 flex justify-center items-center bg-black bg-opacity-75">
-                <iframe
-                  width="70%"
-                  height="70%"
-                  src={`https://www.youtube.com/embed/${movie.trailer}?autoplay=1`}
-                  title={movie.title || movie.name}
-                  frameBorder="0"
-                  allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-                  allowFullScreen
-                ></iframe>
-              </div>
-            )}
           </div>
         ))}
       </Carousel>
