+++ src/services/tmdbApi.js
@@ -38,10 +38,4 @@ export const getPopularTVShows = async () => {
   return response.data.results.slice(0, 10); 
 };
 
-export const getMovieTrailer = async (movieId) => {
-  const response = await tmdbApi.get(`/movie/${movieId}/videos`);
-  const trailer = response.data.results.find(video => video.type === 'Trailer' && video.site === 'YouTube');
-  return trailer ? trailer.key : null;
-};
-
 export default tmdbApi;
